refactor(agents): one file per group of RPC handlers
agent-rpc-handlers.ts was 923 lines, and 690 of them were a single object
literal holding thirteen handlers that share nothing but the logger. Reading
one meant paging past the rest.

Each group is now its own factory, composed back into the same object the
worker calls: the agent config a run reads, conversation persistence, the
flow-step callbacks, tool execution, and sending email. What the groups share,
the run-scoped conversation update and the connection and conversation
lookups, sits in rpc-shared.

Nothing is rewritten. Every line is where it was, in a smaller file: normalise
away imports and export keywords and the new files match the original exactly,
apart from the five braces that close the new factories. Five imports that only
the moved-away handlers used are gone.
